Checkers-Fixed.py
This changes the original print board to a init function, and adds a new function to just print the board. This way, the moves will just be appended to the matrix in the move pieces function, and will print out in the print function. (I still can't get line 31 to work)